Fix and simplify both Raw Classes and Type Aliases
