+++ 07-presentation/src/main/scala/dagmendez/presentation/package.scala
@@ -2,16 +2,16 @@ package dagmendez
 
 package object presentation:
   val controlDigit: Map[Int, String] = Map(
-    0 -> "T",
-    1 -> "R",
-    2 -> "W",
-    3 -> "A",
-    4 -> "G",
-    5 -> "M",
-    6 -> "Y",
-    7 -> "F",
-    8 -> "P",
-    9 -> "D",
+    0  -> "T",
+    1  -> "R",
+    2  -> "W",
+    3  -> "A",
+    4  -> "G",
+    5  -> "M",
+    6  -> "Y",
+    7  -> "F",
+    8  -> "P",
+    9  -> "D",
     10 -> "X",
     11 -> "B",
     12 -> "N",
